Fix a portability issue with u32-vs-i32

+++ crates/jit/src/profiling/jitdump.rs
@@ -55,7 +55,8 @@ impl ProfilingAgent for JitDumpAgent {
         let mut jitdump_file = JITDUMP_FILE.lock().unwrap();
         let jitdump_file = jitdump_file.as_mut().unwrap();
         let timestamp = jitdump_file.get_time_stamp();
-        let tid = rustix::thread::gettid().as_raw_nonzero().get();
+        #[allow(trivial_numeric_casts)]
+        let tid = rustix::thread::gettid().as_raw_nonzero().get() as u32;
         if let Err(err) =
             jitdump_file.dump_code_load_record(&name, addr, size, timestamp, self.pid, tid)
         {
